Switch to Preconstruct for library building
