+++ next-cloudinary/babel.config.js
@@ -1,9 +1,14 @@
 module.exports = api => {
+  const presets = []
+
   if ( api.env('test') ) {
-    return {
-      presets: [['@babel/preset-env', {targets: {node: 'current'}}]],
-    }
+    presets.push([['@babel/preset-env', {targets: {node: 'current'}}]]);
   }
 
-  return {};
+  presets.push(['@babel/preset-react', { runtime: 'automatic' }]);
+  presets.push(['@babel/preset-typescript']);
+
+  return {
+    presets
+  };
 };
