+++ next-cloudinary/src/components/CldOgImage/CldOgImage.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import Head from 'next/head';
 
 import { CldImageProps } from '../CldImage/CldImage';
