+++ next-cloudinary/src/components/CldUploadButton/CldUploadButton.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 import CldUploadWidget, { CldUploadWidgetProps } from '../CldUploadWidget';
