+++ next-cloudinary/src/constants/analytics.ts
