+++ next-cloudinary/src/lib/util.ts
@@ -4,7 +4,7 @@
  */
 
 export function triggerOnIdle(callback: any) {
-  if ( window && 'requestIdleCallback' in window ) {
+  if ( typeof window !== undefined && 'requestIdleCallback' in window ) {
     return requestIdleCallback(callback);
   }
   return setTimeout(() => callback(), 1);
