+++ package.json
@@ -9,6 +9,7 @@
   },
   "devDependencies": {
     "@colbyfayock/semantic-release-pnpm": "^1.2.2",
+    "@preconstruct/cli": "^2.8.7",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
     "semantic-release": "^23.0.0"
